requirements.txt
fastapi
uvicorn
python-multipart
mlflow
transformers
scikit-learn
pandas
tqdm
matplotlib
pillow